Added 2 graphs display

+++ Task 1/Case 2/Neural_Network_2/Matrix.cs
@@ -75,7 +75,7 @@ namespace Matrix
             }
         }
 
-        public void ArrayToMatrix(int[] array, bool bias)
+        public void ArrayToMatrix(double[] array, bool bias)
         {
             for (int i = 0; i < array.Length; ++i)
             {
